Adding authentication for ownership of user info.
New middleware function checks if the email set in the cookies matches
the email for the user info you are trying to update/delete. Will reject
the request if they do not match.

The existing middleware function already matches the token from the RC
API to the email set in the cookies, so the obvious cookie-injection
attack will not work.